requirements.txt
caffe==0.1.0
numpy==2.1.3
opencv_contrib_python==4.10.0.84
Pillow==11.0.0
pyrealsense2==2.55.1.6486
pyvips==2.2.3
timm==1.0.11
torch==2.5.1
torchmetrics==1.4.0.post0
torchvision==0.20.1
tqdm==4.65.2
